Almost Finished
I have updated the settings function and made it so that you can chose what values you want for each character. This allows for ultimate customization of your code. Just  be sure to remember what you changed it to, because it would be extremely hard to figure out the code otherwise.