Allows reading and writing to files
This new update allows reading and writing to a text file, so that you can easily send code to your friends. You can only read and write coded outputs in the file. This is to ensure that you have to know the other's encoding settings in order to successfully decode the message.